Split SubscriptionCache in 4 modules
