+++ src/lib/cache/EntityInfo.cpp
@@ -1,0 +1,82 @@
+/*
+*
+* Copyright 2015 Telefonica Investigacion y Desarrollo, S.A.U
+*
+* This file is part of Orion Context Broker.
+*
+* Orion Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* iot_support at tid dot es
+*
+* Author: Ken Zangelin
+*/
+#include <string>
+#include <semaphore.h>
+
+#include "cache/EntityInfo.h"
+
+
+namespace orion
+{
+
+
+
+/* ****************************************************************************
+*
+* EntityInfo::EntityInfo - 
+*/
+EntityInfo::EntityInfo(const std::string& idPattern, const std::string& _entityType)
+{
+  regcomp(&entityIdPattern, idPattern.c_str(), 0);
+  entityType = _entityType;
+}
+
+
+
+/* ****************************************************************************
+*
+* EntityInfo::match - 
+*/
+bool EntityInfo::match
+(
+  const std::string&  id,
+  const std::string&  type
+)
+{
+  //
+  // If type non-empty - perfect match is mandatory
+  // If type is empty - always matches
+  //
+  if ((type != "") && (entityType != type))
+  {
+    return false;
+  }
+
+  // REGEX-comparison this->entityIdPattern VS id
+  return regexec(&entityIdPattern, id.c_str(), 0, NULL, 0) == 0;
+}
+
+
+
+/* ****************************************************************************
+*
+* EntityInfo::release - 
+*/
+void EntityInfo::release(void)
+{
+  regfree(&entityIdPattern);
+}
+
+}
