+++ src/lib/cache/Subscription.cpp
@@ -1,0 +1,424 @@
+/*
+*
+* Copyright 2015 Telefonica Investigacion y Desarrollo, S.A.U
+*
+* This file is part of Orion Context Broker.
+*
+* Orion Context Broker is free software: you can redistribute it and/or
+* modify it under the terms of the GNU Affero General Public License as
+* published by the Free Software Foundation, either version 3 of the
+* License, or (at your option) any later version.
+*
+* Orion Context Broker is distributed in the hope that it will be useful,
+* but WITHOUT ANY WARRANTY; without even the implied warranty of
+* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero
+* General Public License for more details.
+*
+* You should have received a copy of the GNU Affero General Public License
+* along with Orion Context Broker. If not, see http://www.gnu.org/licenses/.
+*
+* For those usages not covered by this license please contact with
+* iot_support at tid dot es
+*
+* Author: Ken Zangelin
+*/
+#include <string>
+#include <vector>
+
+#include "common/globals.h"
+#include "ngsi10/SubscribeContextRequest.h"
+#include "ngsi10/UpdateContextSubscriptionRequest.h"
+#include "cache/Subscription.h"
+
+
+
+namespace orion
+{
+
+/* ****************************************************************************
+*
+* Subscription::Subscription - 
+*/
+Subscription::Subscription()
+{
+  tenant                = "";
+  servicePath           = "";
+  subscriptionId        = "";
+  throttling            = 0;
+  expirationTime        = 0;
+  lastNotificationTime  = -1;
+  pendingNotifications  = 0;
+  format                = XML;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::Subscription - 
+*/
+Subscription::Subscription
+(
+  const std::string& _tenant,
+  const std::string& _servicePath,
+  const std::string& _subscriptionId,
+  Format             _format
+)
+{
+  tenant                = _tenant;
+  servicePath           = _servicePath;
+  subscriptionId        = _subscriptionId;
+  throttling            = 0;
+  expirationTime        = 0;
+  lastNotificationTime  = -1;
+  pendingNotifications  = 0;
+  format                = _format;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::Subscription - 
+*/
+Subscription::Subscription
+(
+  const std::string&        _tenant,
+  const std::string&        _servicePath,
+  SubscribeContextRequest*  scrP,
+  std::string               _subscriptionId,
+  int64_t                   _expiration,
+  Format                    _format
+)
+{
+  unsigned int ix;
+
+  //
+  // 0. Fill in 'constant initial values'
+  //
+  lastNotificationTime  = -1;
+  pendingNotifications  = 0;
+  format                = XML;
+
+
+
+  //
+  // 1. Get Tenant, ServicePath and Subscription Id
+  //
+  tenant         = _tenant;
+  servicePath    = _servicePath;
+  subscriptionId = _subscriptionId;
+  format         = _format;
+
+  //
+  // 2. Convert all patterned EntityId to EntityInfo
+  //
+  for (ix = 0; ix < scrP->entityIdVector.vec.size(); ++ix)
+  {
+    EntityId* eIdP;
+
+    eIdP = scrP->entityIdVector.vec[ix];
+
+    if (eIdP->isPattern == "false")
+    {
+      continue;
+    }
+
+    EntityInfo* eP = new EntityInfo(eIdP->id, eIdP->type);
+    
+    entityIdInfoAdd(eP);
+  }
+
+
+  //
+  // 3. Insert all attributes 
+  //
+  for (ix = 0; ix < scrP->attributeList.attributeV.size(); ++ix)
+  {
+    attributes.push_back(scrP->attributeList.attributeV[ix]);
+  }
+
+
+  //
+  // 4. throttling
+  //
+  throttling = scrP->throttling.parse();
+
+
+  //
+  // 5. expirationTime
+  //
+  expirationTime = _expiration;
+
+
+  //
+  // 6. restriction
+  //
+  restriction.fill(&scrP->restriction);
+
+
+  //
+  // 7. notifyConditionVector
+  //
+  notifyConditionVector.fill(scrP->notifyConditionVector);
+
+
+  //
+  // 8. reference
+  //
+  reference.set(scrP->reference.get());
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::Subscription - 
+*/
+Subscription::Subscription
+(
+  const std::string&               _tenant,
+  const std::string&               _servicePath,
+  const std::string&               _subscriptionId,
+  const std::vector<EntityInfo*>&  _entityIdInfos,
+  const std::vector<std::string>&  _attributes,
+  int64_t                          _throttling,
+  int64_t                          _expirationTime,
+  const Restriction&               _restriction,
+  NotifyConditionVector&           _notifyConditionVector,
+  const std::string&               _reference,
+  Format                           _format
+)
+{
+  unsigned int ix;
+
+  tenant         = _tenant;
+  servicePath    = _servicePath;
+  subscriptionId = _subscriptionId;
+  format         = _format;
+
+  for (ix = 0; ix < _entityIdInfos.size(); ++ix)
+  {
+    entityIdInfos.push_back(_entityIdInfos[ix]);
+  }
+
+  for (ix = 0; ix < _attributes.size(); ++ix)
+  {
+    attributes.push_back(_attributes[ix]);
+  }
+
+  throttling     = _throttling;
+  expirationTime = _expirationTime;
+
+  restriction.fill((Restriction*) &_restriction);
+  notifyConditionVector.fill(_notifyConditionVector);
+  reference.set(_reference);
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::entityIdInfoAdd - 
+*/
+void Subscription::entityIdInfoAdd(EntityInfo* entityIdInfoP)
+{
+  entityIdInfos.push_back(entityIdInfoP);
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::attributeMatch - 
+*/
+bool Subscription::attributeMatch(const std::string& attributeName)
+{
+  for (unsigned int ncvIx = 0; ncvIx < notifyConditionVector.size(); ++ncvIx)
+  {
+    NotifyCondition* ncP = notifyConditionVector[ncvIx];
+
+    for (unsigned int cvIx = 0; cvIx < ncP->condValueList.size(); ++cvIx)
+    {
+      if (ncP->condValueList[cvIx] == attributeName)
+      {
+        return true;
+      }
+    }
+  }
+
+  return false;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::match - 
+*/
+bool Subscription::match
+(
+  const std::string&  _tenant,
+  const std::string&  _servicePath,
+  const std::string&  id,
+  const std::string&  type,
+  const std::string&  attributeName
+)
+{
+  if (_tenant != tenant)
+  {
+    return false;
+  }
+
+  if (servicePathMatch(_servicePath) == false)
+  {
+    return false;
+  }
+
+
+  //
+  // If ONCHANGE and one of the attribute names in the scope vector
+  // of the subscription has the same name as the incoming attribute. there is a match.
+  //
+  if (!attributeMatch(attributeName))
+  {
+    return false;
+  }
+
+  for (unsigned int ix = 0; ix < entityIdInfos.size(); ++ix)
+  {
+    if (entityIdInfos[ix]->match(id, type))
+    {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::servicePathMatch - 
+*/
+bool Subscription::servicePathMatch
+(
+  const std::string&  _servicePath
+)
+{
+  const char* spath = servicePath.c_str();
+
+  //
+  // No wildcard - exact match
+  //
+  if (spath[strlen(spath) - 1] != '#')
+  {
+    return (_servicePath == servicePath);
+  }
+
+  // Special case. "/#" matches EVERYTHING
+  if (servicePath == "/#")
+  {
+    return true;
+  }
+
+
+  //
+  // Wildcard - remove '#' and compare to the length of _servicePath.
+  //            If equal upto the length of _servicePath, then it is a match
+  //
+  // Actually, there is more than one way to match here:
+  // If we have a servicePath of the subscription as 
+  // "/a/b/#", then the following service-paths must match:
+  //   1. /a/b
+  //   2. /a/b/ AND /a/b/.+  (any path below /a/b/)
+  //
+  // What should NOT match is "/a/b2".
+  //
+  unsigned int len = strlen(spath) - 2;
+
+  // 1. /a/b - (removing 2 chars from the cache-subscription removes "/#"
+
+  if ((spath[len] == '/') && (strlen(_servicePath.c_str()) == len) && (strncmp(spath, _servicePath.c_str(), len) == 0))
+  {
+    return true;
+  }
+
+  // 2. /a/b/.+
+  len = strlen(spath) - 1;
+  if (strncmp(spath, _servicePath.c_str(), len) == 0)
+  {
+    return true;
+  }
+  
+  return false;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::hasAttribute - 
+*/
+bool Subscription::hasAttribute(const std::string& attributeName)
+{
+  unsigned int ix;
+
+  for (ix = 0; ix < attributes.size(); ++ix)
+  {
+    if (attributes[ix] == attributeName)
+    {
+      return true;
+    }
+  }
+
+  return false;
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::update - 
+*/
+void Subscription::update2(UpdateContextSubscriptionRequest* ucsrP)
+{
+  // 1. Update expirationTime if 'duration' is set in UpdateContextSubscriptionRequest
+  if (ucsrP->duration.get() != "")
+  {
+    int64_t _expirationTime = getCurrentTime() + ucsrP->duration.parse();
+    expirationTime  = _expirationTime;
+  }
+
+  // 2. Update throttling if 'throttling' is set in UpdateContextSubscriptionRequest
+  if (ucsrP->throttling.get() != "")
+  {
+    throttling = ucsrP->throttling.parse();
+  }
+
+  // 3. restriction
+  // 4. notifyConditionVector
+}
+
+
+
+/* ****************************************************************************
+*
+* Subscription::release - 
+*/
+void Subscription::release(void)
+{
+  restriction.release();
+  notifyConditionVector.release();
+
+  for (unsigned int ix = 0; ix < entityIdInfos.size(); ++ix)
+  {
+    entityIdInfos[ix]->release();
+    delete(entityIdInfos[ix]);
+  }
+
+  entityIdInfos.clear();
+}
+
+}
