+++ src/lib/mongoBackend/MongoCommonUpdate.cpp
@@ -37,7 +37,9 @@
 #include "common/globals.h"
 #include "common/string.h"
 #include "common/sem.h"
+#include "cache/Subscription.h"
 #include "cache/SubscriptionCache.h"
+#include "cache/subCache.h"
 #include "orionTypes/OrionValueType.h"
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/TriggeredSubscription.h"
